+++ attributes/default.rb
@@ -1,37 +1,38 @@
 #
-# Cookbook Name:: jira
-# Attributes:: default
-#
-# Copyright 2012, SecondMarket Labs, LLC.
-# Copyright 2013, Chef Software, Inc.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
+# JIRA installer configuration
 #
 
-# We need Java 7 for JIRA
-default['java']['jdk_version'] = '7'
+default['jira']['version']  = '6.4.11'
+default['jira']['url']      = "https://www.atlassian.com/software/jira/downloads/binary/atlassian-jira-#{node['jira']['version']}-x64.bin"
+default['jira']['checksum'] = '4030010efd5fbec3735dc3a585cd833af957cf7efe4f4bbc34b17175ff9ba328'
 
-default['jira']['version'] = '6.3.15'
-default['jira']['parentdir'] = '/opt'
+# This will get transformed into a response.varfile for the installer
+default['jira']['install']['rmiPort$Long']                          = '8005'
+default['jira']['install']['app.jiraHome']                          = '/var/atlassian/application-data/jira'
+default['jira']['install']['app.install.service$Boolean']           = 'true'
+default['jira']['install']['existingInstallationDir']               = '/opt/jira'
+default['jira']['install']['sys.confirmedUpdateInstallationString'] = 'false'
+default['jira']['install']['sys.languageId']                        = 'en'
+default['jira']['install']['sys.installationDir']                   = '/opt/atlassian/jira'
+default['jira']['install']['executeLauncherAction$Boolean']         = 'true'
+default['jira']['install']['httpPort$Long']                         = '8080'
+default['jira']['install']['portChoice']                            = 'default'
 
-# This is what the JIRA Installation docs refer to as the "JIRA Installation Directory"
-default['jira']['installdir'] = "#{node['jira']['parentdir']}/atlassian-jira-#{node['jira']['version']}-standalone"
-default['jira']['tarball'] = "atlassian-jira-#{node['jira']['version']}.tar.gz"
-default['jira']['url'] = "http://www.atlassian.com/software/jira/downloads/binary/#{node['jira']['tarball']}"
+# # Additional libraries to add
+# # example: MySQL connector for JIRA
+# default['jira']['lib']['mysql-connector']['source'] = 'http://www.example.com/mysql-connector.jar'
+# default['jira']['lib']['mysql-connector']['path']   = "#{node['jira']['install']['sys.installationDir']}/lib/mysql-connector.jar"
 
-# This is what the JIRA Installation docs refer to as the "JIRA Home Directory"
-default['jira']['homedir'] = '/var/jira-home'
+#
+# JIRA database configuration
+#
 
-default['jira']['crowd_sso']['sso_appname'] = 'jira'
-default['jira']['crowd_sso']['sso_password'] = 'jira'
-default['jira']['crowd_sso']['crowd_base_url'] = 'http://localhost:8095/crowd/'
+default['jira']['dbconfig']['jira-database-config']['name']           = 'defaultDS'
+default['jira']['dbconfig']['jira-database-config']['delegator-name'] = 'default'
+default['jira']['dbconfig']['jira-database-config']['database-type']  = 'hsql'
+default['jira']['dbconfig']['jira-database-config']['schema-name']    = 'public'
+
+default['jira']['dbconfig']['jira-database-config']['jdbc-datasource']['url']          = "jdbc:hsqldb:#{node['jira']['install']['app.jiraHome']}/database/jiradb"
+default['jira']['dbconfig']['jira-database-config']['jdbc-datasource']['driver-class'] = 'org.hsqldb.jdbcDriver'
+default['jira']['dbconfig']['jira-database-config']['jdbc-datasource']['username']     = 'sa'
+default['jira']['dbconfig']['jira-database-config']['jdbc-datasource']['password']     = ''